Add feature to skip the download

+++ process.sh
@@ -10,9 +10,13 @@ else
   exit 1
 fi
 
+# To skip the download, save the "final.osm.pbf" file in the home directory and execute it like this: ./process.sh true
+SKIP_DOWNLOAD=$1
+
 # Delete the tmp directory if exists and create an empty one
 WORK_DIR=$(pwd)
 TMP_DIR=$WORK_DIR/tmp
+SAVE_DIR=/home/christian
 
 if [ -d "$TMP_DIR" ]; then
   rm -rf "$TMP_DIR"
@@ -20,11 +24,17 @@ fi
 
 mkdir -p "$TMP_DIR"
 
-# Download the OSM data
-./download/download.sh $TMP_DIR
+if [ "$SKIP_DOWNLOAD" == "true" ]; then
+  TMP_DIR=$SAVE_DIR
+else
+  # Download the OSM data
+  ./download/download.sh $TMP_DIR
+fi
 
 # Import the OSM data into the PostgreSQL database
 ./import/import.sh $WORK_DIR $TMP_DIR
 
-# Cleanup the tmp directory
-rm -rf "$TMP_DIR"
+if ! [ "$SKIP_DOWNLOAD" == "true" ]; then 
+  # Cleanup the tmp directory
+  rm -rf "$TMP_DIR"
+fi
